Added readbilheader.py
This Python script, readbilheader.py, is designed to extract header information from a BIL (Band Interleaved by Line) file. It parses the content of the corresponding header file associated with the BIL file and extracts relevant information such as the byte order, number of rows and columns, number of bands, nodata value, data type, and other metadata. The extracted information is returned as a dictionary. The script also includes a test section to demonstrate the functionality of the readbilheader function.